+++ packages/expo-calendar/src/Calendar.ts
@@ -374,6 +374,13 @@ export async function updateAttendeeAsync(
   return ExpoCalendar.saveAttendeeForEventAsync(newDetails, null);
 } // Android
 
+export async function getDefaultCalendarAsync(): Promise<Calendar> {
+  if (!ExpoCalendar.getDefaultCalendarAsync) {
+    throw new UnavailabilityError('Calendar', 'getDefaultCalendarAsync');
+  }
+  return ExpoCalendar.getDefaultCalendarAsync();
+} // iOS
+
 export async function deleteAttendeeAsync(id: string): Promise<void> {
   if (!ExpoCalendar.deleteAttendeeAsync) {
     throw new UnavailabilityError('Calendar', 'deleteAttendeeAsync');